feat(event): money top player update event
