+++ build.gradle
@@ -8,7 +8,7 @@ plugins {
 }
 
 group = 'com.nextplugins'
-version = '1.0.7'
+version = '1.1.0'
 
 tasks.build.dependsOn('pdm', 'shadowJar')
