+++ src/main/java/com/nextplugins/economy/listener/events/chat/TycoonTagRegister.java
@@ -1,4 +1,4 @@
-package com.nextplugins.economy.listener.chat;
+package com.nextplugins.economy.listener.events.chat;
 
 import br.com.devpaulo.legendchat.api.events.ChatMessageEvent;
 import com.nextplugins.economy.NextEconomy;
@@ -7,13 +7,14 @@ import com.nextplugins.economy.configuration.values.RankingValue;
 import com.nextplugins.economy.storage.RankingStorage;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
 
 public final class TycoonTagRegister implements Listener {
 
     protected final RankingStorage rankingStorage = NextEconomy.getInstance().getRankingStorage();
 
-    @EventHandler
+    @EventHandler(priority = EventPriority.MONITOR)
     public void onChat(ChatMessageEvent event) {
 
         Player player = event.getSender();
