+++ src/main/java/com/nextplugins/economy/listener/events/operation/MoneyGiveListener.java
@@ -1,4 +1,4 @@
-package com.nextplugins.economy.listener.operation;
+package com.nextplugins.economy.listener.events.operation;
 
 import com.nextplugins.economy.NextEconomy;
 import com.nextplugins.economy.api.event.operations.MoneyGiveEvent;
@@ -10,13 +10,14 @@ import org.bukkit.Bukkit;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
 
 public final class MoneyGiveListener implements Listener {
 
     protected final AccountStorage accountStorage = NextEconomy.getInstance().getAccountStorage();
 
-    @EventHandler
+    @EventHandler(priority = EventPriority.MONITOR)
     public void onDeposit(MoneyGiveEvent event) {
         CommandSender sender = event.getSender();
         Player target = event.getTarget();
