+++ src/main/java/com/nextplugins/economy/listener/events/operation/MoneyWithdrawListener.java
@@ -1,7 +1,7 @@
-package com.nextplugins.economy.listener.operation;
+package com.nextplugins.economy.listener.events.operation;
 
 import com.nextplugins.economy.NextEconomy;
-import com.nextplugins.economy.api.event.operations.MoneySetEvent;
+import com.nextplugins.economy.api.event.operations.MoneyWithdrawEvent;
 import com.nextplugins.economy.api.model.Account;
 import com.nextplugins.economy.configuration.values.MessageValue;
 import com.nextplugins.economy.storage.AccountStorage;
@@ -10,23 +10,24 @@ import org.bukkit.Bukkit;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
 
-public final class MoneySetListener implements Listener {
+public final class MoneyWithdrawListener implements Listener {
 
     protected final AccountStorage accountStorage = NextEconomy.getInstance().getAccountStorage();
 
-    @EventHandler
-    public void onSet(MoneySetEvent event) {
+    @EventHandler(priority = EventPriority.MONITOR)
+    public void onWithdraw(MoneyWithdrawEvent event) {
         CommandSender sender = event.getSender();
         Player target = event.getTarget();
         double amount = event.getAmount();
 
         Account targetAccount = accountStorage.getAccount(target.getUniqueId());
 
-        targetAccount.setBalance(amount);
+        targetAccount.withdrawAmount(amount);
 
-        sender.sendMessage(MessageValue.get(MessageValue::setAmount)
+        sender.sendMessage(MessageValue.get(MessageValue::removeAmount)
                 .replace("$player", Bukkit.getOfflinePlayer(targetAccount.getOwner()).getName())
                 .replace("$amount", NumberUtils.format(amount))
         );
