+++ src/main/java/com/nextplugins/economy/task/AccountRankingTask.java
@@ -1,11 +1,14 @@
 package com.nextplugins.economy.task;
 
 import com.google.common.collect.Lists;
+import com.nextplugins.economy.api.event.operations.MoneyTopPlayerUpdateEvent;
 import com.nextplugins.economy.api.model.Account;
 import com.nextplugins.economy.dao.AccountDAO;
 import com.nextplugins.economy.storage.RankingStorage;
 import lombok.RequiredArgsConstructor;
+import org.bukkit.Bukkit;
 
+import java.time.Instant;
 import java.util.List;
 
 @RequiredArgsConstructor
@@ -20,8 +23,24 @@ public final class AccountRankingTask implements Runnable {
         List<Account> accounts = Lists.newLinkedList(accountDAO.selectAll("ORDER BY balance DESC LIMIT 10"));
 
         if (!accounts.isEmpty()) {
+
+            Account lastAccount = rankingStorage.getRankingAccounts().get(0);
             rankingStorage.getRankingAccounts().clear();
+
             accounts.forEach(rankingStorage.getRankingAccounts()::add);
+
+            Account topAccount = rankingStorage.getRankingAccounts().get(0);
+            if (lastAccount.getOwner().equals(topAccount.getOwner())) return;
+
+            Bukkit.getPluginManager().callEvent(
+                    MoneyTopPlayerUpdateEvent.builder()
+                            .lastMoneyTop(lastAccount)
+                            .moneyTop(topAccount)
+                            .updateInstant(Instant.now())
+                            .async(true)
+                    .build()
+            );
+
         }
 
     }
